requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.38.0
pydantic==2.9.2

# Database
sqlalchemy==2.0.35
sqlite-utils==3.36

# ML / NLP
numpy==1.26.4
scikit-learn==1.5.2

# Vector embeddings
faiss-cpu==1.8.0.post1

# Doc processing
pymupdf==1.24.9    # fitz
pdfplumber==0.11.3

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.1.0
tqdm==4.66.1

# Testing
pytest==8.3.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# HTTP requests
httpx==0.27.2
requests==2.32.3

# Ollama (local LLM client)
ollama==0.1.8  
